% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_interface.R
\name{cv.pros}
\alias{cv.pros}
\title{Cross Validation}
\usage{
cv.pros(X, y, K_fold, alpha, lambdas, algorithm = "subgradient_cd")
}
\arguments{
\item{X}{the data matrix}

\item{y}{the vector response}

\item{K_fold}{partition size}

\item{lambdas}{lambda values to be evaluated}

\item{algorithm}{}
}
\value{
A class \code{cv_pros} with
\itemize{
  \item \code{best_lambda} the  best lambda.
  \item \code{lambdas} the lambda values
  \item \code{risks} the cross-validation risks
}
}
\description{
The K-fold cross-validation function
}
