% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pros.R
\name{predict.cv_pros}
\alias{predict.cv_pros}
\title{CV Prediction}
\usage{
\method{predict}{cv_pros}(cv_prosObj, X_new)
}
\arguments{
\item{cv_prosObj}{an object of class \code{cv_pros}}

\item{X_new}{the data matrix}
}
\value{
A \code{vector} of predictions
}
\description{
The Prediction function
}
