% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pros.R
\name{predict.pros}
\alias{predict.pros}
\title{Pros Prediction}
\usage{
\method{predict}{pros}(object, X, ...)
}
\arguments{
\item{object}{an object of class \code{pros}}

\item{X}{is an \eqn{n \times m}-dimensional matrix of the data.}

\item{...}{Other parameters (this is required by R)}
}
\value{
A \code{vector} of prediction values.
}
\description{
The prediction function for \code{pros}.
}
