% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_interface.R
\name{pros}
\alias{pros}
\title{Fit}
\usage{
pros(X, y, alpha, lambda, algorithm = "subgradient_cd")
}
\arguments{
\item{X}{the data matrix}

\item{y}{the vector response}

\item{alpha}{the vector response}

\item{lambda}{A lambda value}

\item{algorithm}{}
}
\value{
A class \code{pros} with
}
\description{
The fit function for a specific lambda value
}
