% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_interface.R
\name{cv.pros}
\alias{cv.pros}
\title{Cross Validation}
\usage{
cv.pros(X, y, K_fold = 5, alpha = c(1, 0, 0, 0, 0, 0, 0),
  lambdas = seq(10^(-7), 1, 0.1), algorithm = "proximal_gradient_cd",
  max_iter = 10000, tolerance = 10^(-7))
}
\arguments{
\item{X}{the data matrix}

\item{y}{the response vector}

\item{K_fold}{partition size}

\item{alpha}{convex combination}

\item{lambdas}{A vector of Lagrangian penalization values to be evaluated}

\item{algorithm}{the optimization algorithm}
}
\value{
A class \code{cv_pros} with
\itemize{
  \item \code{best_lambda} the  best lambda.
  \item \code{lambdas} the lambda values
  \item \code{risks} the cross-validation risks
}
}
\description{
The K-fold cross-validation function
}
