% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pros.R
\name{cv.pros}
\alias{cv.pros}
\title{Cross-validation}
\usage{
cv.pros(X, y, K_fold = 10, alpha = c(1, 0, 0, 0, 0, 0),
  lambdas = c(), step_size, algorithm = "proximal_gradient_cd",
  max_iter = 10000, tolerance = 10^(-8), random_seed = 0)
}
\arguments{
\item{X}{the matrix of the data}

\item{y}{the vector of response values}

\item{alpha}{the convex combination of length 7 corresponding to the penalties:
\itemize{
  \item l1 penalty
  \item l2 penalty
  \item l4 penalty
  \item l6 penalty
  \item l8 penalty
  \item l10 penalty
}}

\item{lambdas}{A vector of dual penalization values to be evaluated}

\item{step_size}{step size}

\item{algorithm}{the optimization algorithm 
\itemize{
  \item \code{proximal_gradient_cd} (proximal gradient coordinate descent)
  \item \code{subgradient_cd} (subgradient coordinate algorithm)
}}

\item{max_iter}{maximum iterations. This also tunes the step size.}

\item{tolerance}{tolerance}

\item{random_seed}{random seed}
}
\value{
A class \code{cv_pros}
}
\description{
The K-fold cross-validation function.
}
\examples{
cv = cv.pros(X_train, y_train)
pred = predict(cv, X_test)

}
