% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pros.R
\name{cv.pros}
\alias{cv.pros}
\title{Cross-validation}
\usage{
cv.pros(X, y, K_fold = 5, alpha = c(1, 0, 0, 0, 0, 0),
  lambdas = seq(10^(-7), 1, 0.1), algorithm = "proximal_gradient_cd",
  max_iter = 1e+05, tolerance = 10^(-3))
}
\arguments{
\item{X}{the matrix of the data}

\item{y}{the vector of response values}

\item{alpha}{the convex combination of length 7 corresponding to the penalties:
\itemize{
  \item l1 penalty
  \item l2 penalty
  \item l4 penalty
  \item l6 penalty
  \item l8 penalty
  \item l10 penalty
}}

\item{lambdas}{A vector of dual penalization values to be evaluated}

\item{algorithm}{the optimization algorithm 
\itemize{
  \item proximal_gradient_cd
  \item subgradient_cd
}}

\item{max_iter}{maximum iterations. This also tunes the step size.}

\item{tolerance}{tolerance}
}
\value{
A class \code{cv_pros}
}
\description{
The K-fold cross-validation function.
}
\examples{
cv = cv.pros(X_train, y_train)
pred = predict(cv, X_test)

}
