% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pros.R
\name{predict.cv_pros}
\alias{predict.cv_pros}
\title{Cross-validation Prediction}
\usage{
\method{predict}{cv_pros}(object, X_new, ...)
}
\arguments{
\item{object}{an object of class \code{cv_pros}}

\item{X_new}{is an \eqn{n \times m}-dimensional matrix of the data.}

\item{...}{Other parameters (this is required by R)}
}
\value{
A \code{vector} of prediction values.
}
\description{
The prediction function for \code{cv.pros}.
}
