% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_interface.R
\name{predict.pros}
\alias{predict.pros}
\title{Prediction}
\usage{
\method{predict}{pros}(prosObj, X)
}
\arguments{
\item{prosObj}{an object of class \code{pros}}

\item{X}{the data matrix}
}
\value{
A \code{vector} of predictions
}
\description{
The Prediction function
}
