% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pros.R
\name{predict.pros}
\alias{predict.pros}
\title{Pros Prediction}
\usage{
\method{predict}{pros}(prosObj, X)
}
\arguments{
\item{prosObj}{an object of class \code{pros}}

\item{X}{the matrix of the data to predict}
}
\value{
A \code{vector} of prediction values.
}
\description{
The prediction function.
}
\examples{
fit = pros(X_train, y_train, lambda = .1)
pred = predict(fit, X_test)

}
