% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pros.R
\name{pros}
\alias{pros}
\title{Fit}
\usage{
pros(X, y, alpha = c(1, 0, 0, 0, 0, 0, 0), lambda,
  algorithm = "proximal_gradient_cd", max_iter = 10000,
  tolerance = 10^(-7))
}
\arguments{
\item{X}{the data matrix}

\item{y}{the response vector response}

\item{alpha}{convex combination}

\item{lambda}{The Lagrangian penalization value}

\item{algorithm}{the optimization algorithm}
}
\value{
A class \code{pros} with
}
\description{
The fit function for a specific lambda value
}
