% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_interface.R
\name{pros}
\alias{pros}
\title{Fit}
\usage{
pros(X, y, alpha, lambda, algorithm = "proximal_gradient_cd")
}
\arguments{
\item{X}{the data matrix}

\item{y}{the response vector response}

\item{alpha}{convex combination}

\item{lambda}{The Lagrangian penalization value}

\item{algorithm}{the optimization algorithm}
}
\value{
A class \code{pros} with
}
\description{
The fit function for a specific lambda value
}
