% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pros.R
\name{pros}
\alias{pros}
\title{Pros}
\usage{
pros(X, y, alpha = c(1, 0, 0, 0, 0, 0, 0), lambda,
  algorithm = "proximal_gradient_cd", max_iter = 10000,
  tolerance = 10^(-7))
}
\arguments{
\item{X}{the matrix of the data}

\item{y}{the vector of response values}

\item{alpha}{the convex combination of length 7 corresponding to the penalties:
\itemize{
  \item l1 penalty
  \item l2 penalty
  \item l4 penalty
  \item l6 penalty
  \item l8 penalty
  \item l10 penalty
}}

\item{lambda}{the dual penalization value}

\item{algorithm}{the optimization algorithm 
\itemize{
  \item proximal_gradient_cd
  \item subgradient_cd
}}

\item{max_iter}{maximum iterations. This also tunes the step size.}

\item{tolerance}{tolerance}
}
\value{
A class \code{pros}
}
\description{
The fit function for a specific lambda value.
}
\examples{
fit = pros(X_train, y_train, lambda = .1)
pred = predict(fit, X_test)

}
